sistemate le notifiche in background (ancora piccole cose da testare)

+++ app/src/main/java/com/example/firstapp/Alert/ExampleService.java
@@ -1,14 +1,19 @@
 package com.example.firstapp.Alert;
 
 import android.app.Service;
+import android.content.Context;
 import android.content.Intent;
 import android.os.IBinder;
 import android.util.Log;
 import android.widget.EditText;
 import android.widget.TextView;
 import androidx.annotation.Nullable;
+import androidx.room.Room;
+
 import com.example.firstapp.AppDatabase;
 import com.example.firstapp.Channel.Channel;
+
+import java.util.List;
 import java.util.Timer;
 
 /*
@@ -20,9 +25,7 @@ import java.util.Timer;
  * Si dichiara che il programma è in ogni sua parte, opera originale dell'autore
  *
  */
-
 public class ExampleService extends Service {
-
     private static Double tempMin;
     private static Double tempMax;
     private static Double umidMin;
@@ -47,22 +50,43 @@ public class ExampleService extends Service {
     private static Channel channel;
     private static AppDatabase database;
     private static TextView notification;
+    private static Context context;
+    private static boolean go=false;
 
     @Override
     public void onCreate() {
         super.onCreate();
+        context=getApplicationContext();
+        Log.d("ExampleService","servizio background creato");
+        database = Room.databaseBuilder(getApplicationContext(), AppDatabase.class, "prodiction")
+                //consente l'aggiunta di richieste nel thred principale
+                .allowMainThreadQueries()
+                .fallbackToDestructiveMigration()
+                //build mi serve per costruire il tutto
+                .build();
     }
 
     //funzione che devo fare all'avvio
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
-
-        timerTask = new MyTimerTask(url,channel, temp,
-                umid,  ph,  cond,  irra,  peso);
-        timer = new Timer();
-        timer.scheduleAtFixedRate(timerTask, 0, 3000);
-
-        return START_REDELIVER_INTENT;
+            //recupero la lista del channel attuale
+            String id = database.SavedDao().getAll().get(0).getId();
+            String key = database.SavedDao().getAll().get(0).getKey();
+            channel = database.ChannelDao().findByName(id, key);
+        if(channel.getNotification()) {
+            database.ChannelDao().delete(channel);
+            url = "https://api.thingspeak.com/channels/" + channel.getId() + "/feeds.json?api_key=" + channel.getRead_key() + "&results=1";
+            channel.setTimerTask(new MyTimerTask(url, channel, temp, umid, ph, cond, irra, peso, context));
+            timer = new Timer();
+            timer.scheduleAtFixedRate(channel.getTimerTask(), 0, 3000);
+            channel.setTimer(timer);
+            database.ChannelDao().insert(channel);
+        }
+        else{
+            this.onDestroy();
+            return  START_NOT_STICKY;
+        }
+        return START_STICKY;
     }
 
     public static void setvalue(EditText tempMin1, EditText tempMax1, EditText umidMin1, EditText umidMax1, EditText condMin1, EditText condMax1, EditText phMin1, EditText phMax1, EditText irraMin1,
@@ -73,8 +97,8 @@ public class ExampleService extends Service {
         notification=notifiche;
         url=url1;
         try {
-        tempMin = Double.valueOf(tempMin1.getText().toString());
-        channel.setTempMin(tempMin);
+             tempMin = Double.valueOf(tempMin1.getText().toString());
+             channel.setTempMin(tempMin);
         } catch (NumberFormatException e) {
             tempMin = null;
         }
@@ -171,24 +195,26 @@ public class ExampleService extends Service {
     }
 
     public static void stoptimer(){
-        if(timer!=null){
-            timer.cancel();
-            Channel x=database.ChannelDao().findByName(channel.getId(),channel.getRead_key());
-            database.ChannelDao().delete(x);
-            x.setNotification(false);
-            database.ChannelDao().insert(x);
-            notification.setText("notifiche non attive");
-            Log.d("Background service","Servizio interrotto!");
+        if(channel!=null) {
+            timer = channel.getTimer();
+            timerTask = channel.getTimerTask();
+            if (timer != null) {
+                timer.cancel();
+                timerTask.cancel();
+                Channel x = database.ChannelDao().findByName(channel.getId(), channel.getRead_key());
+                database.ChannelDao().delete(x);
+                x.setNotification(false);
+                x.setTimer(null);
+                x.setTimerTask(null);
+                database.ChannelDao().insert(x);
+                if (notification != null) notification.setText("notifiche non attive");
+                Log.d("Background service", "Servizio interrotto!");
+            }
         }
-        if(timerTask!=null) timerTask.cancel();
-
-
     }
 
     @Override
-    public void onDestroy() {
-        super.onDestroy();
-    }
+    public void onDestroy() { super.onDestroy(); stoptimer(); Log.d("ExampleServices","distruggo");}
 
     @Nullable
     @Override
